Add ObjectDrivable
