+++ source/game/field/CourseColMgr.cc
@@ -574,25 +574,6 @@ bool CourseColMgr::doCheckMaskOnlyPush(KColData *data, CollisionCheckFunc collis
     return hasCol;
 }
 
-void CourseColMgr::CollisionInfo::update(f32 now_dist, const EGG::Vector3f &offset,
-        const EGG::Vector3f &fnrm, u32 kclAttributeTypeBit) {
-    bbox.min = bbox.min.minimize(offset);
-    bbox.max = bbox.max.maximize(offset);
-
-    if (kclAttributeTypeBit & KCL_TYPE_FLOOR) {
-        updateFloor(now_dist, fnrm);
-    } else if (kclAttributeTypeBit & KCL_TYPE_WALL) {
-        if (wallDist > -std::numeric_limits<f32>::min()) {
-            f32 dot = 1.0f - wallNrm.ps_dot(fnrm);
-            if (dot > perpendicularity) {
-                perpendicularity = std::min(dot, 1.0f);
-            }
-        }
-
-        updateWall(now_dist, fnrm);
-    }
-}
-
 CourseColMgr *CourseColMgr::s_instance = nullptr; ///< @addr{0x809C3C10}
 
 } // namespace Field
