+++ source/game/field/CourseColMgr.hh
@@ -17,40 +17,6 @@ typedef bool (
 /// @nosubgrouping
 class CourseColMgr : EGG::Disposer {
 public:
-    struct CollisionInfoPartial {
-        EGG::BoundBox3f bbox;
-        EGG::Vector3f tangentOff;
-    };
-
-    struct CollisionInfo {
-        EGG::BoundBox3f bbox;
-        EGG::Vector3f tangentOff;
-        EGG::Vector3f floorNrm;
-        EGG::Vector3f wallNrm;
-        EGG::Vector3f _3c;
-        f32 floorDist;
-        f32 wallDist;
-        f32 _50;
-        f32 perpendicularity;
-
-        void updateFloor(f32 dist, const EGG::Vector3f &fnrm) {
-            if (dist > floorDist) {
-                floorDist = dist;
-                floorNrm = fnrm;
-            }
-        }
-
-        void updateWall(f32 dist, const EGG::Vector3f &fnrm) {
-            if (dist > wallDist) {
-                wallDist = dist;
-                wallNrm = fnrm;
-            }
-        }
-
-        void update(f32 now_dist, const EGG::Vector3f &offset, const EGG::Vector3f &fnrm,
-                u32 kclAttributeTypeBit);
-    };
-
     struct NoBounceWallColInfo {
         EGG::BoundBox3f bbox;
         EGG::Vector3f tangentOff;
