+++ source/game/field/KColData.cc
@@ -682,4 +682,23 @@ KColData::KCollisionPrism::KCollisionPrism(f32 height, u16 posIndex, u16 faceNor
     : height(height), pos_i(posIndex), fnrm_i(faceNormIndex), enrm1_i(edge1NormIndex),
       enrm2_i(edge2NormIndex), enrm3_i(edge3NormIndex), attribute(attribute) {}
 
+void CollisionInfo::update(f32 now_dist, const EGG::Vector3f &offset, const EGG::Vector3f &fnrm,
+        u32 kclAttributeTypeBit) {
+    bbox.min = bbox.min.minimize(offset);
+    bbox.max = bbox.max.maximize(offset);
+
+    if (kclAttributeTypeBit & KCL_TYPE_FLOOR) {
+        updateFloor(now_dist, fnrm);
+    } else if (kclAttributeTypeBit & KCL_TYPE_WALL) {
+        if (wallDist > -std::numeric_limits<f32>::min()) {
+            f32 dot = 1.0f - wallNrm.ps_dot(fnrm);
+            if (dot > perpendicularity) {
+                perpendicularity = std::min(dot, 1.0f);
+            }
+        }
+
+        updateWall(now_dist, fnrm);
+    }
+}
+
 } // namespace Field
