+++ source/game/kart/KartMove.hh
@@ -84,7 +84,7 @@ public:
     void calcHopPhysics();
     void calcRejectRoad();
     bool calcZipperCollision(f32 radius, f32 scale, EGG::Vector3f &pos, EGG::Vector3f &upLocal,
-            const EGG::Vector3f &prevPos, Field::CourseColMgr::CollisionInfo *colInfo,
+            const EGG::Vector3f &prevPos, Field::CollisionInfo *colInfo,
             Field::KCLTypeMask *maskOut, Field::KCLTypeMask flags) const;
     f32 calcSlerpRate(f32 scale, const EGG::Quatf &from, const EGG::Quatf &to) const;
     virtual void calcVehicleRotation(f32 turn);
